requirements.txt
azure-common==1.1.24
azure-storage-common==2.1.0
azure-storage-file==2.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
fastapi==0.49.0
h11==0.9.0
httptools==0.1.1
idna==2.9
pycparser==2.19
pydantic==1.4
python-dateutil==2.8.1
python-dotenv==0.11.0
requests==2.23.0
git+https://github.com/swisscom-blockchain/sbca-indy-wrapper@1.8.1
six==1.14.0
starlette==0.12.9
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
